+++ BBGE/DebugFont.cpp
@@ -89,7 +89,7 @@ void DebugFont::formatText()
 	text = this->text;
 	lines.clear();
 	std::string currentLine;
-	int lastSpace = -1;
+	size_t lastSpace = size_t(-1);
 	float currentWidth = 0;
 	maxW = 0;
 	for (size_t i = 0; i < text.size(); i++)
@@ -103,7 +103,7 @@ void DebugFont::formatText()
 				lastSpace = i;
 			}
 			lines.push_back(text.substr(0, lastSpace));
-			int tsz = text.size();
+			size_t tsz = text.size();
 			text = text.substr(lastSpace+1, tsz);
 			i = 0;
 			maxW = std::max(maxW, currentWidth);
